requirements.txt
addict==2.4.0
aliyun-python-sdk-core==2.15.0
aliyun-python-sdk-kms==2.16.2
av==8.0.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.1.1
crcmod==1.7
cryptography==42.0.5
cycler==0.12.1
Cython==3.0.9
fonttools==4.50.0
idna==3.6
importlib_metadata==7.1.0
importlib_resources==6.4.0
Jinja2==3.1.3
jmespath==0.10.0
kiwisolver==1.4.5
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.7.5
mdurl==0.1.2
mmcv==2.1.0
mmengine==0.10.3
model-index==0.1.11
mpld3==0.5.10
numpy==1.24.4
opencv-python==4.9.0.80
opendatalab==0.0.10
openmim==0.3.9
openxlab==0.0.37
ordered-set==4.1.0
oss2==2.17.0
packaging==24.0
pandas==2.0.3
pillow==10.2.0
PIMS==0.5
platformdirs==4.2.0
pycocotools==2.0.7
pycparser==2.21
pycryptodome==3.20.0
Pygments==2.17.2
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2023.4
PyYAML==6.0.1
requests==2.28.2
rich==13.4.2
scipy==1.10.1
shapely==2.0.3
six==1.16.0
slicerator==1.1.0
tabulate==0.9.0
termcolor==2.4.0
terminaltables==3.1.10
tomli==2.0.1
tqdm==4.61.1
typing_extensions==4.10.0
tzdata==2024.1
urllib3==1.26.18
yapf==0.40.2
zipp==3.18.1
